Add Day 02 logic and tests
Added a new method `Day02` in `TestData.cs` to provide test data for Day 02.
Introduced a new test class `Day02Tests` in `Tests.cs` to define expected results for Day 02.
Created a new class `Day02` in `Day02.cs` that extends the `Day` class and implements the logic for processing parts one and two of Day 02.
The `Day02` class includes methods to:
  * Count valid lines in `ProcessPartOne`.
  * Calculate a "safe" value by iterating through reports and skipping invalid lines in `ProcessPartTwo`.
  * Determine if a line is valid based on specific criteria in the `ValidLine` method.